More detailed logging when the token validation fails
In case the token is invalid the client only gets a generic error.
During setting up the stack it would be nice if there was a way
to debug potential errors, so log the causing exception by default.

This includes for example if the issues ID or time is the cause of the error.