Port to web-token/jwt-library
Since we don't use the deprecated functions and the namespaces
stayed the same, there is nothing to port.

It now emits a deprecation warning when no clock is passed, so
use the symfony clock implementation there. symfony/clock is
new in symfony 6 but should work fine with symfony 5 too, since
there are no dependencies.